+++ src/ui/utButton.hpp
@@ -1,5 +1,4 @@
 #pragma once
-#include <Geode/Geode.hpp>
 using namespace geode::prelude;
 
 class utButton: public CCMenuItemSpriteExtra {