+++ src/ui/utMenuBackground.hpp
@@ -1,7 +1,5 @@
 #pragma once
-#include <Geode/Geode.hpp>
 using namespace geode::prelude;
-
 
 class utMenuBackground : public CCLayer {
 public: